requirements.txt
streamlit
pandas
gspread
google-auth
oauth2client
google-api-python-client